= Version 3.01 = * Invoice currency settings fixed. * Bug with loading of custom templates fixed. * Added an ability to use manual payment method if there is no any payment venue accepted. * Fixed bug with PayPal IPN which logged wrong amount. * Fixed bug with email notifications. * Default template layout and design improvements. * Other simple fixes and improvements.
git-svn-id: https://plugins.svn.wordpress.org/wp-invoice/trunk@429233 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ core/gateways/class_wpi_paypal.php
@@ -71,7 +71,7 @@ class wpi_paypal extends wpi_gateway_base {
         case 'Completed':
           // Add payment amount
           $event_note = WPI_Functions::currency_format(abs($_POST['mc_gross']), $_POST['invoice'])." paid via PayPal";
-          $event_amount = (int)$_POST['mc_gross'];
+          $event_amount = (float)$_POST['mc_gross'];
           $event_type   = 'add_payment';
 
           // Log balance changes
